+++ src/Integration/Monolog/CloudWatch/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- Added support for `async-aws/cloud-watch-logs` 2.0
+
 ## 1.0.2
 
 ### Changed